fix: build bug
